feat: refactoring after comments
