+++ segment_spotify others.go
@@ -1,0 +1,8 @@
+// +build !darwin
+// +build !windows
+
+package main
+
+func (s *spotify) enabled() bool {
+	return false
+}
